requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pyyaml>=6.0
mwparserfromhell>=0.6.6
python-dateutil>=2.8.2
